refactor: Renamed main file
